Add necessary static_cast
The number of tabs isn't going to be great enough where the distinction
between something like int and size_t matters.

+++ Explorer++/Explorer++/TabHandler.cpp
@@ -1585,7 +1585,7 @@ boost::optional<Tab> Explorerplusplus::GetTab(int tabId)
 
 int Explorerplusplus::GetNumTabs() const
 {
-	return m_Tabs.size();
+	return static_cast<int>(m_Tabs.size());
 }
 
 int Explorerplusplus::MoveTab(const Tab &tab, int newIndex)
